move sublime lxml related methods to separate file
